some github-inspired tweaks
